backend/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask-Cors==5.0.0
marqo==3.8.1
python-dotenv==1.0.1
Requests==2.32.3
llama_cpp_python==0.3.1
black
nltk==3.9.1